Allow floats for distance calcuation rather than doing a ceil()

+++ GalCon-Client/GalCon-Client/src/com/xxx/galcon/model/Move.java
@@ -4,7 +4,7 @@ public class Move {
 	public String fromPlanet;
 	public String toPlanet;
 	public int shipsToMove = 0;
-	public int duration = 0;
+	public float duration = 0;
 	public String playerHandle;
 	
 	public boolean belongsToPlayer(Player player){
